pyproject.toml
[project]
name = "toosh_co_il"
version = "0.0.1"
dependencies = ["flask", "Pillow"]
requires-python = ">=3.12"
authors = [{ name = "Cutipus", email = "Cutipus@proton.me" }]
description = "Backend for toosh.co.il website"
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[project.urls]
Repository = ""
Issues = ""
